Move 'map_nodes' to graph_utils.py
